chore(type): change declarations of `ClientConfig` and `SongPiece`
This commit includes two major changes.

The first one is the new field `root` of `ClientConfig`. This interface
is to store `client`-side of the **parsed** configurations. The original
three fields can give the real root URL of the front-end, but it's not
very easy and necessary to calculate it every time. So we add this field
which is updated when `parseConfig()` is called so that we can directly
get this value.

The second one is the changed types of `cover_art_url` (which is a new
name of `cover` field) and `lrc` fields. `SongPiece` means the **raw**
properties of one song - for both front-end and back-end. The back-end
returns null values for cover and lrc when necessary, so we should
prepare for this edge case. As for the change of the field name, the
reason is similar: the back-end returns `cover_art_url` instead of
`cover`, so we must change it.